requirements.txt
-r core/requirements.txt

alabaster==0.7.12
backports.functools-lru-cache==1.6.1
backports.ssl-match-hostname==3.7.0.1
beautifulsoup4==4.9.1
cffi==1.14.3
cryptography==3.1
# defusedxml is required for SAML authentication
defusedxml==0.6.0
dm.xmlsec.binding==1.3.7
# docker is used for the LCP feature.
docker==4.3.1
docutils==0.16
expiringdict==1.2.1
Flask-Cors==3.0.9
flask-sqlalchemy-session==1.1
# freezegun is used for unit tests.
freezegun==0.3.15
html-sanitizer==1.9.1
httplib2==0.18.1
imagesize==1.2.0
# ndg-httpsclient ensures we support SNI-based SSL
ndg-httpsclient==0.5.1
# newrelic is an NYPL-sepcific requirement that might be removeable.
newrelic==5.18.0.148
oauth2client==4.1.3
packaging==20.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.8
# pyfakefs is used for unit tests.
pyfakefs==3.7
Pygments==2.5.2
PyJWT==1.4.2
PyLD==0.7.3
pypostalcode==0.3.4
pyOpenSSL==19.1.0
# TODO: python-Levenshtein is supposedly required for author name
# matching, but is not a core requirement; perhaps it can be removed.
python-Levenshtein==0.12.0
# python-saml is required for SAML authentication
python-saml==2.8.0
# requests-mock is used for unit tests.
requests-mock==1.8.0
# rsa is an oauth2client dependency. 4.3 is the last version to support Python 2.
rsa==4.3
snowballstemmer==2.0.0
soupsieve==1.9.6
# Sphinx is for code documentation
Sphinx==1.8.5
sphinxcontrib-websupport==1.1.2
typing==3.7.4.3
unicodecsv==0.14.1
uritemplate==3.0.1
uszipcode==0.2.4
wcag-contrast-ratio==0.9
websocket-client==0.57.0
# xmlsec is required for SAML authentication
xmlsec==1.3.8
alchemy-mock==0.4.3
